+++ crates/pallets/starknet/runtime_api/src/lib.rs
@@ -17,7 +17,7 @@ pub extern crate alloc;
 use alloc::vec::Vec;
 
 use mp_simulations::{
-    FeeEstimate, InternalSubstrateError, SimulationError, SimulationFlags, TransactionSimulationResult,
+    FeeEstimate, InternalSubstrateError, SimulationError, SimulationFlags, TransactionSimulationResult, ReExecutionResult
 };
 use mp_starknet_inherent::L1GasPrices;
 use sp_api::BlockT;
@@ -25,8 +25,6 @@ use starknet_api::core::{ClassHash, ContractAddress, EntryPointSelector, Nonce};
 use starknet_api::hash::StarkFelt;
 use starknet_api::state::StorageKey;
 use starknet_api::transaction::{Calldata, Event as StarknetEvent, MessageToL1, TransactionHash};
-
-type ReExecutionResult = Result<Vec<(TransactionExecutionInfo, Option<CommitmentStateDiff>)>, SimulationError>;
 
 sp_api::decl_runtime_apis! {
     pub trait StarknetRuntimeApi {
@@ -75,6 +73,8 @@ sp_api::decl_runtime_apis! {
         /// Idealy, the execution traces of all of `transactions_to_trace`.
         /// If any of the transactions (from both arguments) fails, an error is returned.
         fn re_execute_transactions(transactions_before: Vec<Transaction>, transactions_to_trace: Vec<Transaction>, with_state_diff: bool) -> Result<ReExecutionResult, InternalSubstrateError>;
+
+        fn get_transaction_re_execution_state_diff(transactions_before: Vec<Transaction>, transactions: Vec<Transaction>) -> Result<Result<CommitmentStateDiff, SimulationError>, InternalSubstrateError>;
 
         fn get_index_and_tx_for_tx_hash(xts: Vec<<Block as BlockT>::Extrinsic>, tx_hash: TransactionHash) -> Option<(u32, Transaction)>;