+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - feat: actual estimate_fee added, brought back l1 messages and refactored
   simulate tx
+- dev: impl get_state_updates using get_transaction_re_execution_state_diff
 - feat: support strk as fee token
 - dev: pallet test for estimate_fee that skip validation
 - feat: add versioned constants to pallet constants